Add files via upload
Correlation analysis and visualization through heat maps and scatter plots: This notebook presents a detailed analysis of correlation between variables using coefficients such as Pearson, Kendall and Spearman. In addition, heat maps are generated to visualize relationships and patterns between variables. Practical examples are provided and the steps required to load data, calculate correlation coefficients and create heat maps are explained. It is a useful tool for exploring and understanding the relationships between variables in multidimensional data sets.